// src/api/common/seeders/enums/seeder.enum.ts
export enum SeederEnum {
  CRM_ACCOUNTS = 'CRM_ACCOUNTS',
  CRM_LOCATIONS = 'CRM_LOCATIONS',
  MIGRATIONS = 'MIGRATIONS',
  OC_NCE = 'OC_NCE',
  ROLE_PERMISSIONS = 'ROLE_PERMISSIONS',
  SESSIONS_FILTERS = 'SESSIONS_FILTERS',
  STAFFING_MANAGEMENT_STAFF_LIST_FILTERS = 'STAFFING_MANAGEMENT_STAFF_LIST_FILTERS',
  STAFFING_MANAGEMENT_STAFF_SCHEDULE_FILTERS = 'STAFFING_MANAGEMENT_STAFF_SCHEDULE_FILTERS',
  OC_APPROVALS = 'OC_APPROVALS',
  OC_DRIVES = 'OC_DRIVES',
  STAFFING_MANAGEMENT_SCHEDULE_DETAILS_NOTES_FILTERS = 'STAFFING_MANAGEMENT_SCHEDULE_DETAILS_NOTES_FILTERS',
  CRM_CONTACTS_DONORS = 'CRM_CONTACTS_DONORS',
}